% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDLSMOTE.R
\name{BDLSMOTE}
\alias{BDLSMOTE}
\title{The borderline-SMOTE algorithm.}
\usage{
BDLSMOTE(data, perc_min = 50, perc_over = NULL, over_replace = FALSE,
  k = 5, m = 2 * (k + 1), borderline = 1, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{perc_min}{The desired \% size of the minority class relative to the
whole data set. For instance, if \code{perc_min} = 50 the returned data
set is balanced. \code{perc_min} is ignored if \code{perc_over} is
specified.}

\item{perc_over}{\% of examples to append to the input data set relative
to the size of the minority class. For instance, if \code{perc_over} = 100
the minority class doubles in size. If specified, \code{perc_min} is
ignored.}

\item{over_replace}{A logical value indicating whether the neighbours 
picked from the \code{k} nearest neighbours should be picked with or 
without replacement.}

\item{k}{Number of nearest neighbours to compute for each example in the 
minority class.}

\item{m}{Number of neighbours used to decide whether a minority example 
is borderline or not. The authors call the set of all borderline examples 
the DANGER set.}

\item{borderline}{Select between borderline-SMOTE1 and borderline-SMOTE2. 
Possible values are 1 and 2.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a more balanced version of the input data 
 after over-sampling with either borderline-SMOTE1 or borderline-SMOTE2.
}
\description{
\code{BDLSMOTE} over-samples the input data using either borderline-SMOTE1 
or borderline-SMOTE2 algorithms.
}
\details{
Borderline-SMOTE\{1, 2\} algorithms work similarly to SMOTE, however, they 
only synthesise new examples using the minority examples that are 
borderline.
}
\references{
Han, H., Wang, W. Y., & Mao, B. H. (2005, August).
 Borderline-SMOTE: a new over-sampling method in imbalanced data sets
 learning. In \emph{International Conference on Intelligent Computing}
 (pp. 878-887). Springer Berlin Heidelberg.
}
