% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENN.R
\name{ENN}
\alias{ENN}
\title{The Edited Nearest Neighbours algorithm.}
\usage{
ENN(data, remove_class = "Majority", k = 3, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{remove_class}{Examples from \code{remove_class} are removed. The 
options are: \code{c("Majority", "Minority", "Both")}.}

\item{k}{Number of nearest neighbours to take into account.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a clean version of the input data set after 
 application of the Edited Nearest Neighbours algorithm.
}
\description{
\code{ENN} removes the examples that are misclassified by their \code{k} 
nearest neighbours.
}
\details{
The default behaviour of \code{ENN} is to remove examples from the majority 
class that are misclassified by their \code{k} nearest neighbours, however, 
the user can modify this and select to remove only examples from the 
minority class or remove examples from both classes.
}
\references{
Wilson, D. L. (1972). Asymptotic properties of nearest 
neighbor rules using edited data. \emph{IEEE Transactions on Systems, Man, 
and Cybernetics}, \emph{2}(3), 408-421.
}
