% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_imbalanced_data}
\alias{generate_imbalanced_data}
\title{Generate an imbalanced data set.}
\usage{
generate_imbalanced_data(num_examples = 100L, num_features = 2L,
  imbalance_ratio = 5, noise_maj = 0.05, noise_min = 0.1, seed = NULL)
}
\arguments{
\item{num_examples}{Total number of examples in the data set.}

\item{num_features}{Total number of features in the data set.}

\item{imbalance_ratio}{Ratio of the number of examples in the majority 
class to the number of examples in the minority class.}

\item{noise_maj}{Fraction of the minority class that is mislabelled as 
majority class.}

\item{noise_min}{Fraction of the majority class that is mislabelled as 
minority class.}

\item{seed}{Integer value for reproducibility purposes.}
}
\value{
A data frame containing an imbalanced two-class data set.
}
\description{
\code{generate_imbalanced_data} is a simple function to generate a 
two-class imbalanced data set.
}
\details{
The imbalanced data set generated has two classes where the majority class 
comes from a multivariate normal distribution with mean zero and unitary 
standard deviation for all features and the minority class comes from a 
multivariate normal distribution with mean two and unitary standard 
deviation for all features.

The total number of examples and the dimensionality of the data are chosen 
through the \code{num_examples} and \code{num_features} arguments. The 
\code{imbalance_ratio} argument together with \code{num_examples} 
determines the exact number of examples in the majority and minority 
classes. To simulate noise in the data, approximately \code{noise_min} 
examples in the majority class are labelled as minority class examples and 
approximately \code{noise_maj} examples in the minority class are labelled 
as majority class examples. \code{noise_maj} and \code{noise_min} are 
fractions.
}
