% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.R
\name{identify_tomek_links}
\alias{identify_tomek_links}
\title{Identify Tomek Links.}
\usage{
identify_tomek_links(data)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}
}
\value{
A logical vector indicating for each example in the input data set 
 whether it belongs to a Tomek Link or not.
}
\description{
\code{identify_tomek_links} returns a vector indicating whether each 
example in the input data set belong to a Tomek Link or not.
}
