% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWMOTE.R
\name{MWMOTE}
\alias{MWMOTE}
\title{The MWMOTE algorithm.}
\usage{
MWMOTE(data, perc_min = 50, perc_over = NULL, k1 = 5, k2 = 3,
  k3 = round(nrow(data_min_filtered)/2), cut_off = 5, max_closeness = 2,
  cluster_complexity = 3, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{perc_min}{The desired \% size of the minority class relative to the
whole data set. For instance, if \code{perc_min} = 50 the returned data
set is balanced. \code{perc_min} is ignored if \code{perc_over} is
specified.}

\item{perc_over}{\% of examples to append to the input data set relative
to the size of the minority class. For instance, if \code{perc_over} = 100
the minority class doubles in size. If specified, \code{perc_min} is
ignored.}

\item{k1}{Number of neighbours used to identify noisy minority examples.}

\item{k2}{Number of neighbours used to identify the borderline majority 
examples.}

\item{k3}{Number of neighbours used to identify the informative minority 
examples.}

\item{cut_off}{Cut-off value to compute the closeness factor.}

\item{max_closeness}{Maximum value for the closeness factor.}

\item{cluster_complexity}{Value utilised to tune the trade-off between the 
number of clusters and their size. A large value leads to larger clusters 
but fewer of them, whereas a small value leads to smaller clusters but 
more of them.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a more balanced version of the input data 
 after over-sampling with the MWMOTE algorithm.
}
\description{
\code{MWMOTE} over-samples the input data using the Majority Weighted 
Over-Sampling TEchnique.
}
\details{
MWMOTE is a complex over-sampling algorithm and comprises three main 
phases. First, the hard-to-learn minority examples are identified, then 
an importance weight is assigned to each of the hard-to-learn examples, and 
finally new examples are synthesised following a strategy similar to SMOTE.

For clarity, the hyperparameters Cf(th), CMAX, and Cp in the original 
description of MWMOTE were renamed here to \code{cut_off}, 
\code{max_closeness}, and \code{cluster_complexity}, respectively.
}
\references{
Barua, S., Islam, M. M., Yao, X., & Murase, K. (2014). 
 MWMOTE--majority weighted minority oversampling technique for imbalanced 
 data set learning. \emph{IEEE Transactions on Knowledge and Data 
 Engineering}, \emph{26}(2), 405-425.
}
