% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCL.R
\name{NCL}
\alias{NCL}
\title{The Neighbourhood Cleaning Rule algorithm.}
\usage{
NCL(data, k = 3, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{k}{Number of nearest neighbours to take into account.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a clean version of the input data set after 
 application of the Neighbourhood Cleaning Rule algorithm.
}
\description{
\code{NCL} removes examples from majority class that are either 
misclassified by their \code{k} nearest neighbours or contributed to the 
misclassification of examples from minority class.
}
\references{
Laurikkala, J. (2001, July). Improving identification of 
 difficult small classes by balancing class distribution. In 
 \emph{Conference on Artificial Intelligence in Medicine in Europe} (pp. 
 63-66). Springer, Berlin, Heidelberg.
}
