% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUS.R
\name{RUS}
\alias{RUS}
\title{The Random Under-Sampling algorithm.}
\usage{
RUS(data, perc_maj = 50, perc_under = NULL, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The
outcome must be both a binary valued factor and the last column of
\code{data}.}

\item{perc_maj}{The desired \% size of the majority class relative to the
whole data set. For instance, if \code{perc_maj} = 50 a balanced version
of the input data set is returned. \code{perc_maj} is ignored if
\code{perc_under} is specified.}

\item{perc_under}{\% of examples to select from the majority class. If
specified \code{perc_maj} is ignored.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a more balanced version of the input data
 set after application of the Random Under-Sampling algorithm. The original
 order of the examples is preserved.
}
\description{
\code{RUS} returns a more balanced version of a data set after application
of the Random Under-Sampling algorithm.
}
\details{
The Random Under-Sampling algorithm creates a new data set containing all 
examples from the minority class plus a random selection of examples from 
the majority class.
}
\examples{
imb_data <- generate_imbalanced_data(num_examples = 200, 
                                     num_features = 2,
                                     imbalance_ratio = 5,
                                     noise_maj = 0,
                                     noise_min = 0,
                                     seed = 42)
 
table(imb_data$target)
table(RUS(imb_data, perc_maj = 50)$target)    # Balance the classes
table(RUS(imb_data, perc_under = 20)$target)  # Select 20\% of maj. class
}
