% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWO.R
\name{RWO}
\alias{RWO}
\title{The Random Walk Over-Sampling algorithm.}
\usage{
RWO(data, perc_min = 50, perc_over = NULL, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{perc_min}{The desired \% size of the minority class relative to the
whole data set. For instance, if \code{perc_min} = 50 the returned data
set is balanced. \code{perc_min} is ignored if \code{perc_over} is
specified.}

\item{perc_over}{\% of examples to append to the input data set relative
to the size of the minority class. For instance, if \code{perc_over} = 100
the minority class doubles in size. If specified, \code{perc_min} is
ignored.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a more balanced version of the input data 
 after over-sampling with the Random Walk Over-Sampling algorithm.
}
\description{
\code{RWO} over-samples the input data using the Random Walk Over-Sampling 
algorithm.
}
\details{
Random Walk Over-Sampling is based on the Central Limit Theorem and 
synthesises new examples for the minority class by perturbing the 
available minority examples.
}
\references{
Zhang, H., & Li, M. (2014). RWO-Sampling: A random walk
 over-sampling approach to imbalanced data classification.
 \emph{Information Fusion}, \emph{20}, 99-116.
}
