% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_sequence.R
\name{sampling_sequence}
\alias{sampling_sequence}
\title{Chain sampling algorithms together.}
\usage{
sampling_sequence(data, algorithms, parameters = NULL, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. 
Restrictions about the input data depend on the individual sampling 
algorithms being used. In any case, the outcome must be both a binary 
valued factor and the last column of \code{data}.}

\item{algorithms}{A vector containing the names of the sampling algorithms 
to be chained either as strings or the functions' objects themselves.}

\item{parameters}{A list of lists where each individual list contains the 
parameters to be used by the respective sampling algorithm in 
\code{algorithms}.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing the result of a chain of sampling 
 algorithms applied to the input data set.
}
\description{
\code{sampling_sequence} applies a sequence of sampling algorithms to an 
input data set.
}
\details{
\code{sampling_sequence} has two main arguments: \code{algorithms} and 
\code{parameters}. \code{algorithms} is a vector containing the names of 
the sampling functions to be applied in sequence (either as strings or 
the functions' objects themselves). \code{parameters} is a list of lists, 
where each individual list contains the parameters to be used by the 
respective sampling function in \code{algorithms}.
}
