% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBC.R
\name{SBC}
\alias{SBC}
\title{The Under-Sampling Based on Clustering algorithm.}
\usage{
SBC(data, perc_maj = 50, perc_under = NULL, k = 3, max_iter = 100L,
  nstart = 10L, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{perc_maj}{The desired \% size of the majority class relative to the
whole data set. For instance, if \code{perc_maj} = 50 a balanced version
of the input data set is returned. \code{perc_maj} is ignored if
\code{perc_under} is specified.}

\item{perc_under}{\% of examples to select from the majority class. If
specified \code{perc_maj} is ignored.}

\item{k}{Number of clusters for the \code{k}-Means algorithm.}

\item{max_iter}{Maximum number of iterations of the k-Means algorithm.}

\item{nstart}{Number of random restarts of the k-Means algorithm.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a more balanced version of the input data 
 after under-sampling with the Under-Sampling Based on Clustering algorithm.
}
\description{
\code{SBC} under-samples the input data using the Under-Sampling Based on 
Clustering algorithm.
}
\details{
Under-Sampling Based on Clustering clusters the input data into \emph{k} 
clusters and randomly selects a number of majority examples from each 
cluster based on the imbalance ratio of the cluster.

The authors did not specify if sampling of majority examples should be 
performed with or without replacement, however, in many occasions the 
algorithm tries to sample more examples than what is available in the 
cluster, therefore, we always perform sampling with replacement here.
}
\references{
Yen, S. J., & Lee, Y. S. (2009). Cluster-based under-sampling
 approaches for imbalanced data distributions.
 \emph{Expert Systems with Applications}, \emph{36}(3), 5718-5727.
}
