% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTE.R
\name{SMOTE}
\alias{SMOTE}
\title{The SMOTE algorithm.}
\usage{
SMOTE(data, perc_min = 50, perc_over = NULL, k = 5,
  over_replace = FALSE, classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{perc_min}{The desired \% size of the minority class relative to the
whole data set. For instance, if \code{perc_min} = 50 the returned data
set is balanced. \code{perc_min} is ignored if \code{perc_over} is
specified.}

\item{perc_over}{\% of examples to append to the input data set relative
to the size of the minority class. For instance, if \code{perc_over} = 100
the minority class doubles in size. If specified, \code{perc_min} is
ignored.}

\item{k}{Number of nearest neighbours to compute for each example in the 
minority class.}

\item{over_replace}{A logical value indicating whether the neighbours 
picked from the \code{k} nearest neighbours should be picked with or 
without replacement.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a more balanced version of the input data 
 set after application of the SMOTE algorithm.
}
\description{
\code{SMOTE} returns a more balanced version of a data set after 
application of the SMOTE algorithm.
}
\details{
SMOTE is an over-sampling algorithm that synthesises new examples in the 
line segment joining two close minority class examples.
}
\references{
Chawla, N. V., Bowyer, K. W., Hall, L. O., & Kegelmeyer, W. P.
 (2002). SMOTE: synthetic minority over-sampling technique. \emph{Journal
 of artificial intelligence research}, \emph{16}, 321-357.
}
