% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.R
\name{TL}
\alias{TL}
\title{Tomek Links.}
\usage{
TL(data, remove_class = "Majority", classes = NULL)
}
\arguments{
\item{data}{A data frame containing the predictors and the outcome. The 
predictors must be numeric and the outcome must be both a binary valued 
factor and the last column of \code{data}.}

\item{remove_class}{Examples from \code{remove_class} are removed. The 
options are: \code{c("Majority", "Minority", "Both")}.}

\item{classes}{A named vector identifying the majority and the minority
classes. The names must be "Majority" and "Minority". This argument is
only useful if the function is called inside another sampling function.}
}
\value{
A data frame containing a clean version of the input data set 
 after removing examples that belong to Tomek Links.
}
\description{
\code{TL} is a cleaning algorithm that removes examples that belong to 
Tomek Links. A pair of examples form a Tomek Link if they belong to 
different classes and are the nearest neighbours of each other.
}
\details{
The user has control over the examples that are removed and can select to 
remove only examples from majority class (under-sampling), minority class, 
and from both classes (cleaning).
}
\references{
Tomek, I. (1976). An experiment with the edited 
nearest-neighbor rule. \emph{IEEE Transactions on systems, Man, and 
Cybernetics}, (6), 448-452.
}
